unlimited certified copies in a domain

+++ libraries/entities/src/EntityTree.cpp
@@ -42,6 +42,7 @@
 
 static const quint64 DELETED_ENTITIES_EXTRA_USECS_TO_CONSIDER = USECS_PER_MSEC * 50;
 const float EntityTree::DEFAULT_MAX_TMP_ENTITY_LIFETIME = 60 * 60; // 1 hour
+static const QString DOMAIN_UNLIMITED = "domainUnlimited";
 
 EntityTree::EntityTree(bool shouldReaverage) :
     Octree(shouldReaverage)
@@ -300,7 +301,7 @@ void EntityTree::postAddEntity(EntityItemPointer entity) {
 
         // Delete an already-existing entity from the tree if it has the same
         //     CertificateID as the entity we're trying to add.
-        if (!existingEntityItemID.isNull()) {
+        if (!existingEntityItemID.isNull() && !entity->getCertificateType().contains(DOMAIN_UNLIMITED)) {
             qCDebug(entities) << "Certificate ID" << certID << "already exists on entity with ID"
                 << existingEntityItemID << ". Deleting existing entity.";
             deleteEntity(existingEntityItemID, true);
@@ -1870,7 +1871,7 @@ int EntityTree::processEditPacketData(ReceivedMessage& message, const unsigned c
                         failedAdd = true;
                         qCDebug(entities) << "User without 'certified rez rights' [" << senderNode->getUUID()
                             << "] attempted to add a certified entity with ID:" << entityItemID;
-                    } else if (isClone && isCertified) {
+                    } else if (isClone && isCertified && !properties.getCertificateType().contains(DOMAIN_UNLIMITED)) {
                         failedAdd = true;
                         qCDebug(entities) << "User attempted to clone certified entity from entity ID:" << entityIDToClone;
                     } else if (isClone && !isCloneable) {
